move converters to core level
